requirements.txt
torch
torchvision
torchaudio
numpy
scikit-learn
PyYAML
tqdm
matplotlib